refactor(react): share server store construction between SSR adapters
createServerStore builds the redux-style store, controller and manager
wiring once so the Express and Next.js adapters stop duplicating it.
